power_manager: Add infrastructure to work with udev tags
We introduce a new infrastructure for matching devices, where udev rules
can set tags on a device, which are then accessible to powerd. This
allows us to use the udev rules language instead of implementing custom
matching functionality. We also provide a simpler way to handle devices
that are already present on startup.

udev rules can add tags to the variable POWERD_TAGS which are then
available from within powerd. Additionally, udev rules should include
"powerd" in the udev TAGS variable to indicate that powerd should
receive events for this device.

BUG=chromium:399007
TEST=Run unit tests.
TEST=Deploy on device along with udev rules (run udevadm trigger after
TEST=changing udev rules). Restart powerd with --vmodule=udev=1 and look
TEST=at the log.  Check that input devices are picked up on startup;
TEST=add/remove external mouse and check that the corresponding events
TEST=are logged.

Change-Id: I3b7ef3be635242db3eb10ffb81fdced42e5dc417
Signed-off-by: Patrik Fimml <patrikf@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/212667