shill: Add check for IsActive on service before disconnect
DisconnectFrom is called when WiFi services are being unloaded.
In this function, a warning message is logged even in the case of
a legitimate state.  This CL adds a wrapper function to DisconnectFrom
to first check if the service is active prior to initiating a
disconnect.  If it is not active, a message is logged.  If it is active,
DisconnectFrom is called and the warning message will be triggered in
the unlikely event of a service being current or pending while also not
active.  Added tests to exercise this functionality.

BUG=chromium:397649
TEST=Ran shill unit tests and validated in net log when disabling wifi
on peppy

Change-Id: I18ee587ed180a8a57f0a59d0ca2bf7786f4e2374
Reviewed-on: https://chromium-review.googlesource.com/212342
Tested-by: Rebecca Silberstein <silberst@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Rebecca Silberstein <silberst@chromium.org>