power_manager: Grant permissions for power/wakeup on all devices
For USB input devices, wakeup properties can only be set on the parent
USB device, which is not classified as an input device itself. This
property is currently written by the usb-hid-wake.sh script which runs
as root.

In preparation for merging this functionality into powerd, we extend the
udev rule to grant permissions to powerd for all power/wakeup files.

BUG=chromium:399007
TEST=tested on peppy

Change-Id: I34f8ca9cb072e63112596b60843907e5a3baeb56
Signed-off-by: Patrik Fimml <patrikf@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/212721
Reviewed-by: Daniel Erat <derat@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>