common-mk: support running arbitrary commands
The current --bin support is limited in that you can't pass any flags
to the program you want to run.  Rework the command line so that all
unused options get passed through to the program.  So now you can do:
	./platform2_test.py --sysroot /build/link /bin/bash
	./platform2_test.py --sysroot /build/link -- /bin/bash -l

The --bin flag is retained both for backwards compatibility, and to
make it easier to set argv[0] when testing.  For example:
	./platform2_test.py --sysroot /build/link --bin /bin/bash -- sh
This will execute the /bin/bash ELF, but with argv[0]="sh".  Some
programs do argv[0] probing, so it's nice if we allowed power users
to poke that.  This doesn't complicate the interface otherwise.

BUG=chromium:401876
TEST=see desc above for commands tested
TEST=`cbuildbot amd64-generic-full` passes

Change-Id: Ia07fa3386261443c69003c31dade48accd55a987
Reviewed-on: https://chromium-review.googlesource.com/212867
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>