installer: Remove vblock copying code.
The vblock as installed with the kernel is now correct and we don't have to
copy in the one from from the stateful partition. Eliminate this now
unnecessary work and remove a dependence on vmlinuz_hd.vblock.

BUG=chromium:403031
TEST=Built test and recovery images for lumpy and booted each. Ran
chromeos-install manually for the test image and rebooted, verifying that the
installed kernel was bootable. Let chromeos-install run automatically for the
recovery image and rebooted, verifying that the installed kernel was bootable.
Ran a tryjob for butterfly-paladin with --hwtest.

Change-Id: I9cc6d4d196bc1003a72ce3aac0e415e5f973643b
Reviewed-on: https://chromium-review.googlesource.com/213287
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>