quipper: Read just magic and size to determine file mode
Previously, a full perf_file_header was read, even if the file was a
pipe mode file. This would leave unknown values in header_ from the file.
Instead, read into piped_header_, and then re-read header_ if necessary.
header_ and piped_header_ are in a union, but for simplicity, the full
header_ is re-read.

TEST=unit tests. A later commit's unit test will require this.
BUG=none

Change-Id: Ib4157aaa630ad0266c1a66d3cb0b897e608507e8
Reviewed-on: https://chromium-review.googlesource.com/214113
Reviewed-by: Emma Rapati <rapati@chromium.org>
Tested-by: David Sharp <dhsharp@google.com>
Commit-Queue: David Sharp <dhsharp@google.com>