shill: Remove suspend timeout action from PowerManager client object.
shill currently posts a timeout when it receives a SuspendImminent message from
power manager. If it doesn't receive a matching SuspendDone within the given
time, it fakes a SuspendDone message.

The original reason for this timeout was that powerd did not guarantee the
SuspendDone message, in case the suspend request was canceled. This is no longer
the case.

This behaviour is a problem because we may incorrectly resume shill due to time
spent in dark resume. This CL removes the timeout.

BUG=chromium:407786
TEST=(1) Run shill unittests.
     (2) Verify suspend-resume re-enables wifi.
     (3) Verify shill resumes after a canceled suspend. To simulate:
         $ powerd_dbus_suspend -delay 2 -wakeup_timeout 15 &
         $ sleep 2.5
         $ killall -9 powerd

Change-Id: I2a127f20e90a193a5f1e1f2fdcbb8d3bb7237371
Reviewed-on: https://chromium-review.googlesource.com/214590
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>