common-mk: avoid --board when possible
The only reason we use --board is to run portageq-$BOARD, and that's
to get settings from the environment.  When running inside of the
portage env, we already have those vars exported, so we don't need
to call portageq at all.

Stop making --board mandatory when the vars we want are already
exported.  This lets the ebuilds stop passing --board, but if the
env var we need isn't available, we'll still get errors rather
than go silently.  This shouldn't happen normally though, so lets
see where this goes wrong as-is.

BUG=chromium:407731
TEST=`cbuildbot x86-generic-full` passes

Change-Id: Ia14030b60ed3279fde6e21857ea96e0a6d4fecf9
Reviewed-on: https://chromium-review.googlesource.com/214690
Reviewed-by: Bertrand Simonnet <bsimonnet@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>