common-mk: workaround kernel binfmt register limits
The kernel places a limit of 256 bytes on the string used to register
the binfmt.  When all the bytes are escaped, we quickly exceed that.
So let's decode the escape sequences like the kernel would, and then
send binary data to the kernel.

However, there's a caveat.  The kernel cannot handle embedded NUL
characters, so we have to manually escape those back to \x00.

However, there's yet another problem.  Even if we only escape the
NUL bytes, we still try to write more than 256 bytes to the kernel
(which fails).  So employ a trick where we modify the magic on the
fly (when possible) to be a non-NUL byte.

This requires tweaking of the MaskMatches function to better match
the behavior of the kernel.

BUG=chromium:401876
TEST=registering mipsel32 no longer breaks 32bit x86

Change-Id: I3870901ce2ba252a75fa0d9d3cc62020282039f9
Reviewed-on: https://chromium-review.googlesource.com/214882
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>