libchromeos: Add http::Transport::CreateDefault to hide CURL
HTTP utility functions in libchromeos require a Transport which
is currently implemented using libcurl. In attempt to hide the
explicit dependency on curl::Transport, I created a CreateDefault
method which now creates a default transport (currently the curl).
But if we want to switch from curl to another transport, it will
be a localized change and not spread out throughout code base.

BUG=None
TEST=FEATURES=test emerge-link libchromeos

Change-Id: I44526f8034a9a53f104a7fd1cb97e822cc2157c8
Reviewed-on: https://chromium-review.googlesource.com/216475
Tested-by: Alex Vakulenko <avakulenko@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Alex Vakulenko <avakulenko@chromium.org>