power_manager: Use inhibit when available
In our kernels, we introduce the "inhibit" feature to selectively
disable input devices that might provide bogus input. We make powerd
use this when it is available.

Also, we only set the "wakeup" and "inhibit" tags on the input* devices,
not on the event* devices.

BUG=chromium:391046
TEST=Tested on link (easy to reproduce bogus input due to metal case).
1. Deploy, udevadm trigger, restart powerd. Make sure that you are
running a kernel that doesn't have our kernel-level lid-filter hacks
applied and has inhibit support instead.
2. With powerd stopped and evtest, verify that bogus input is seen from
touchpad/touchscreen when lid is closed.
3. Start powerd. No bogus input when lid is closed, suspend works
correctly.

Change-Id: I39fb83ea45e70954d54d64f158fcd0a3d3716343
Reviewed-on: https://chromium-review.googlesource.com/217385
Commit-Queue: Patrik Fimml <patrikf@chromium.org>
Tested-by: Patrik Fimml <patrikf@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>