quipper: Change events_ to vector<unique_ptr<event_t>>
Simplifies memory management.

Because event_t are allocated with malloc/calloc, unique_ptr must be given a
deleter that uses free(). To do this, introduce a type alias,
malloced_unique_ptr<T>, into utils.h, and use this for smart pointers to
event_t.

Also improves a possible stack smashing in PerfSerializer::DeserializeEvent via
buffer overrun caused by assuming an event_t has enough space. This is not
necessarily the case because event_t members have flexible array members, which
are not included in sizeof(event_t), and the sample id data are placed in the
memory after the event_t. We probably got away with it thus far because some
members of the union are fairly large. Now, malloc an over-sized array, and
realloc it later. After the fact, CHECK that the initial size was sufficient.
If it wasn't, then we overran the array, and it is not safe to continue.

To correctly fix this, it would be necessary to predict the final size of the
event.

TEST=unit tests
BUG=none

Change-Id: I3906c18892caf96f6d9f7acc2e44e3d14592379f
Reviewed-on: https://chromium-review.googlesource.com/217128
Reviewed-by: Emma Rapati <rapati@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Commit-Queue: David Sharp <dhsharp@chromium.org>
Tested-by: David Sharp <dhsharp@chromium.org>