shill: wifi: remove NIC wake on disconnect programming
Comment out code that programs the NIC to wake on disconnect
before we go into suspend. The reason for this is that disconnection
from an AP while sleeping will cause a full system resume. Only
put this programming back in after the kernel/powerd is capable
of identifying a wake on disconnect signal and putting the system
into dark resume to handle it.

BUG=chromium:413407
TEST='P2_TEST_FILTER="shill::*" FEATURES="test" USE="clang asan"
emerge-squawks platform2' succeeds.

Change-Id: I3bd06b669fa5c330e519ffc2ee66a828c832c992
Reviewed-on: https://chromium-review.googlesource.com/217783
Tested-by: Samuel Tan <samueltan@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Samuel Tan <samueltan@chromium.org>