power_manager: Inhibit early upon lid close
When we receive a lid close event, the first thing we want to do is
inhibit input devices so that we can prevent spurious input events. We
pull WakeupController to the front of this sequence, since
StateController can potentially take a long time to process a lid event.

BUG=chromium:414485
TEST=Tested on link. With an external display attached and e.g. evtest,
verify that input devices are inhibited fast enough to prevent spurious
input events.

Change-Id: If06bf2411daa80295183e6c3a698d28d18efa90b
Reviewed-on: https://chromium-review.googlesource.com/218298
Reviewed-by: Daniel Erat <derat@chromium.org>
Commit-Queue: Patrik Fimml <patrikf@chromium.org>
Tested-by: Patrik Fimml <patrikf@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>