installer: Improve log message when FindKernelConfig() fails.
The previous error message collides with one of the error messages
FindKernelConfig() can print, making it unclear if the failure was
printed from DumpKernelConfig() in the installer code or from
FindKernelConfig() in vboot_host code. Also, the previous error
message might be incorrect if the failure was not "parsing" the
kernel config.

BUG=chromium:415050
TEST=emerge-link chromeos-base/chromeos-installer

Change-Id: Ib477fb592f7e7fa49a44af7d54d8c07ebe77a022
Reviewed-on: https://chromium-review.googlesource.com/218639
Commit-Queue: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>