login_manager: ui: Make dark_resume_always be owned by chronos
In order to determine whether the system woke up due to network activity
or due to user activity, we need a firmware change to expose a mechanism
to query this information.  This change is definitely not going to make
it into the M39 branch but we still need some way to get basic lucid
sleep functionality working by then.

As a temporary workaround, we have exposed a sysfs entry on systems with
kernel 3.14.  When this file is set to 1, the kernel treats all resumes
as dark resumes and relies on chrome and powerd to detect user activity
and transition to fully resumed if necessary.  Chrome will only set this
entry to 1 if the wake-on-packets chrome flag is set to true but in
order to do so, the file must be owned by chronos.

BUG=chromium:414949
CQ-DEPEND=CL:218930
TEST=manual
Verified that the file is owned by chronos.

Change-Id: I90979bca49a5569108a272af451c9421cf041b2b
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/219078
Reviewed-by: Daniel Erat <derat@chromium.org>