buffet: Added error handling to PropType::CreateValue()
PropType::CreateValue(Any) used to abort if Any didn't contain
the data of expected type. Now that we are adding more and more
reliance on D-Bus transport for GCD data (commands and state),
we can no longer guarantee that the variant data passed to
CreateValue contains the expected values (since it may be not
under our control). Add the ability for CreateValue to gracefully
fail and provide necessary error information to the caller.

BUG=chromium:415364
TEST=FEATURES=test emerge-link buffet

Change-Id: I097d674073fc84a69184a78edf2b7381608c000f
Reviewed-on: https://chromium-review.googlesource.com/219135
Tested-by: Alex Vakulenko <avakulenko@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Alex Vakulenko <avakulenko@chromium.org>