common-mk: handle the / sysroot
When running unittests in /, we don't want to strip off the leading /
path as it creates relative paths everywhere.  The code assumes that
this func returns absolute paths when given absolute paths, so just
skip the strip when dealing with this path.

To make sure other weird values don't mess us up, normalize the path
the user gave us with the same algorithm chromite uses for paths.

BUG=chromium:401876
TEST=`FEATURES=test emerge update_engine` passes

Change-Id: Ia248e105a848bc2ac15f853db5b6e86d1b342320
Reviewed-on: https://chromium-review.googlesource.com/219276
Reviewed-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>