installer: Get rid of the skip_vblock option.
We no longer have to muck around with vblocks when installing an image, and
the skip_vblock option no longer does anything. Code search also shows that we
don't use that option internally, so we should be able to get rid of it.

BUG=None
TEST=Built a lumpy image and ran chromeos-install to install to it off of a
USB key.

Change-Id: I22caa84751a3430f2d67e28f34fb71dce76a8e59
Reviewed-on: https://chromium-review.googlesource.com/220701
Reviewed-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Reviewed-by: Bowgo Tsai <bowgotsai@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>