shill: add metrics identifiying the source of wifi services
Services can be configured in the default (system-wide) profile,
or in user-specific profiles. We'd like to know how different types
of WiFi services are configured.

For unencrypted networks, we expect that all will be in the default
profile. This is because Chrome always configures such networks into
the default profile.

For encrypted networks, the answer is less clear. The answer depends
on whether the network was set up before login (in which case it will
be in the default profile), or after login (in which case the logged
in user can choose whether the network is in the default profile, or
the user's profile).

Add a couple of new metrics, to help us figure out what the world
looks like.

While there: move the remembered network count reporting code to
using std::count_if.

BUG=chromium:418723
TEST=unit tests (on peppy)

Change-Id: I3a6bc7324e13d2fa6bcb659b500b9a7c85222913
Reviewed-on: https://chromium-review.googlesource.com/220671
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>