libchromeos: Add support for OUT parameters in D-Bus method invoker
Added support for OUT parameters in D-Bus method handlers. Now it
is possible to return more than one value from a D-Bus call.

All OUT parameters are specified as pointer types in method argument
list. The types must be default-constructible since D-Bus
serialization layer will need to create instances of objects to
provide pointers for when invoking the handler.

When the handler returns, all the values set in the pointed-to
data will be sent back in method response message.

If the handler both has OUT parameters and returns a non-void value,
the OUT parameters will be sent back first before the return
value (that is, the non-void return value is treated as an
implicit OUT parameter at the end of handler's argument list).

BUG=chromium:419271
TEST=FEATURES=test emerge-link libchromeos

Change-Id: Id47c3fa6b2bd2ba754d8e528375a3b3de3c891c8
Reviewed-on: https://chromium-review.googlesource.com/221325
Reviewed-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Alex Vakulenko <avakulenko@chromium.org>
Tested-by: Alex Vakulenko <avakulenko@chromium.org>